chore: move the Hash structure to the interface
