+++ games/src/ataxx/mod.rs
@@ -2,7 +2,6 @@
 // modules public, so they can be accessed
 // without their parent namespace.
 pub use self::bitboard::*;
-pub use self::hash::*;
 pub use self::piece::*;
 pub use self::position::*;
 pub use self::r#move::*;
@@ -10,7 +9,6 @@ pub use self::square::*;
 
 // Non-namespaced modules.
 mod bitboard;
-mod hash;
 mod r#move;
 mod piece;
 mod position;
